+++ CHANGELOG.md
@@ -4,7 +4,7 @@
   - Added support to Enclosure group actions
   - Added support for Enclosure `:refresh` and `:reconfigure`
   - Added support to Ethernet network `:bulk_create`
-  - Added support to Logical Enclosure actions
+  - Added support to Logical Enclosure `:reconfigure`, `:set_script`
 
 ### 0.1.1
   - Fixed Ruby SDK version to 1.0.0